scripts: make vendored scripts also use $TMPDIR/scripts-mkswu
This simplifies external scripts looking at mkswu internal state by
using the same scripts-mkswu directory for state

The drawback is that we must not remove $MKSWU_TMP when running
the pre script from SWU after the installed one was run, so
this duplicates a bit of code, but there should be no drawback
after this.

This also flags that we are using vendored scripts in a new 'vendored'
file, so post rootfs can decide to re-exec itself if required.

+++ scripts_pre.sh
@@ -11,29 +11,53 @@ MKSWU_TMP="$TMPDIR/scripts-mkswu"
 # SCRIPTSDIR is overridden for scripts embedded with swupdate
 SCRIPTSDIR="$MKSWU_TMP"
 
-if [ "${SWUPDATE_CHAIN_IDX:-1}" = 1 ]; then
-	# only re-create dir for first swu in chain
-	rm -rf "$MKSWU_TMP"
-	mkdir "$MKSWU_TMP" || exit 1
-	# 2026/02: also remove old scripts dir for leftovers of older SWUs
-	# (prints error message if missing so recreate)
-	rm -rf "$TMPDIR/scripts"
-	mkdir -p "$TMPDIR/scripts"
+if [ -e "$TMPDIR/sw-description" ]; then
+        SWDESC="$TMPDIR/sw-description"
+elif [ -e "/var/tmp/sw-description" ]; then
+        SWDESC="/var/tmp/sw-description"
+elif [ -e "/tmp/sw-description" ]; then
+        SWDESC="/tmp/sw-description"
 else
-	if ! [ -e "$MKSWU_TMP" ]; then
-		echo "Chained SWU but $MKSWU_TMP does not exist!" >&2
-		exit 1
-	fi
-	# cleanup any certs and re-extract (they are installed
-	# after each individual swu if present)
-	# Note scripts-mkswu is hardcoded as we only consider the
-	# SWU tarball.
-	# 2026/02: also remove old scripts dir for leftovers of older SWUs
-	rm -rf "$TMPDIR/scripts-mkswu/certs_"* \
-		"$TMPDIR/scripts/certs_"*
+        echo "sw-description not found!" >&2
+	exit 1
 fi
 
+# only cleanup state directories once, if either:
+# - running vendored (SCRIPTSDIR changed, runs first)
+# - running embedded version and vendored was not run (sw-description not marked for skip)
+if [ "$SCRIPTSDIR" != "$MKSWU_TMP" ] \
+  || ! grep -q "DEBUG_SKIP_SCRIPTS" "$SWDESC"; then
+	if [ "${SWUPDATE_CHAIN_IDX:-1}" = 1 ]; then
+		# only re-create dir for first swu in chain,
+		rm -rf "$MKSWU_TMP"
+		# 2026/02: also remove old scripts dir for leftovers of older SWUs
+		# (prints error message if missing so recreate)
+		rm -rf "$TMPDIR/scripts"
+		mkdir -p "$TMPDIR/scripts"
+	else
+		if ! [ -e "$MKSWU_TMP" ]; then
+			echo "Chained SWU but $MKSWU_TMP does not exist!" >&2
+			exit 1
+		fi
+		# cleanup any certs and re-extract (they are installed
+		# after each individual swu if present)
+		# 2026/02: also remove old scripts dir for leftovers of older SWUs
+		rm -rf "$MKSWU_TMP/certs_"* \
+			"$TMPDIR/scripts/certs_"*
+	fi
+fi
+
+mkdir -p "$MKSWU_TMP" || exit 1
 cd "$MKSWU_TMP" || exit 1
+
+# remember if we're running vendored or not for post_rootfs script
+# (vendored has a different SCRIPTSDIR)
+if [ "$SCRIPTSDIR" != "$MKSWU_TMP" ]; then
+	touch "vendored"
+else
+	# also remove it for chained update alternating vendored or not...
+	rm -f "vendored"
+fi
 
 # extract archive after script
 sed -e '1,/^BEGIN_ARCHIVE/d' "$0" | tar xv || exit 1
